ARTEMIS-3261 Updating logic to use only replaceable records on compacting verification
(cherry picked from commit 0edf599adca9365993579a96d717ed2f3c0d0af9)

downstream: ENTMQBR-4831

Note: This commit was actually originally developed as part of 2.16.0.jbossorg-x and then cherry-picked into upstream and rebased there.

+++ artemis-journal/src/main/java/org/apache/activemq/artemis/core/journal/impl/FileWrapperJournal.java
@@ -59,7 +59,7 @@ public final class FileWrapperJournal extends JournalBase {
    protected volatile JournalFile currentFile;
 
    @Override
-   public void replaceableRecord(int recordType) {
+   public void replaceableRecord(byte recordType) {
       journal.replaceableRecord(recordType);
    }
